Add calendar script to seed script
